feat: update payment page to use live data
